Change upgradestep order of contactfolder migration.
Because reindexing the contactfolder objects after class migration tries already to adapt the ITranslatedTitle behavior, we had to change the upgradestep order.
