+++ opengever/contact/upgrades/profiles/4600/types/opengever.contact.contactfolder.xml
@@ -2,6 +2,9 @@
 <object xmlns:i18n="http://xml.zope.org/namespaces/i18n"
         name="opengever.contact.contactfolder" meta_type="Dexterity FTI">
 
-  <property name="klass">opengever.contact.contactfolder.ContactFolder</property>
+  <!-- enabled behaviors -->
+  <property name="behaviors" purge="False">
+    <element value="opengever.base.behaviors.translated_title.ITranslatedTitle" />
+  </property>
 
 </object>
